+++ src/main/java/com/vonage/client/voice/DtmfPayload.java
@@ -26,6 +26,12 @@ class DtmfPayload extends JsonableBaseObject {
     @JsonIgnore final String uuid;
     private final String digits;
 
+    /**
+     * Create a new DtmfPayload object.
+     *
+     * @param digits DTMF digits to send to the call as a string.
+     * @param uuid TUUID of the call to play DTMF into.
+     */
     public DtmfPayload(String digits, String uuid) {
         if ((this.digits = digits) == null || digits.trim().isEmpty()) {
             throw new IllegalArgumentException("Must include at least one digit to send.");