+++ CHANGELOG.md
@@ -6,6 +6,9 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 - Added user/domain name support to `com.vonage.client.voice.ncco.SipEndpoint`
 - Improved Voice API documentation
 - Deprecated primitive wrapper methods in NCCO action builders
+- Refactored `com.vonage.client.voice.CallsFilter` to extend `HalFilterRequest`
+  - Deprecated `java.util.Date`-based methods
+  - `getOrder()` returns a `com.vonage.client.common.SortOrder` instead of `CallOrder`
 
 # [8.16.2] - 2025-02-05
 - Added `disconnected_by` enum to `com.vonage.client.voice.EventWebhook`